zen: Patch hid-lenovo driver to be compatible with my patched firmware
At work I have the Lenovo USB keyboard with Trackpoint, which doesn't
send normal button events for the middle mouse button: instead it
tries to emulate mouse wheel events and doesn't do a great job at
it. I've patched the firmware in mine to not do this and just let the
middle button be a regular button, like it is on Thinkpads, but the
kernel driver expects the broken behaviour, so I have to patch it too.